Move XML load / parse methods to Parser class.
